Handle case in _configure_openmp() where MacOS system doesn't have Homebrew
On a MacOS system that's not using Homebrew, running `brew --prefix llvm`
with `subprocess.check_output` doesn't produce a CalledProcessError; instead,
when the shell produces a FileNotFoundError it passes it back and
`check_output` raises that instead.

If we check for OSError (the parent of FileNotFoundError, which should handle
other OS-related errors that could get passed back) as well, then this should
correctly handle this case.

The MacPorts check doesn't have the same issue, because the `which` command
is built in and so `which port` will raise a CalledProcessError if MacPorts
isn't installed.